requirements.txt
# ── Web Framework ──────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# ── Agent Orchestration ────────────────────────────────────────────────────────
langgraph>=0.1.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0

# ── LLM / Embeddings ──────────────────────────────────────────────────────────
openai>=1.25.0
# Uncomment to use DeepSeek instead:
# openai>=1.25.0   # DeepSeek uses the OpenAI-compatible SDK

# ── RAG / Vector Search ────────────────────────────────────────────────────────
llama-index-core>=0.10.0
llama-index-vector-stores-chroma>=0.1.0
llama-index-embeddings-openai>=0.1.0
chromadb>=0.5.0

# ── GitHub Integration ─────────────────────────────────────────────────────────
PyGithub>=2.3.0

# ── Git Operations ─────────────────────────────────────────────────────────────
gitpython>=3.1.40

# ── Config / Validation ────────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# ── HTTP / Utilities ───────────────────────────────────────────────────────────
httpx>=0.27.0
python-multipart>=0.0.9

# ── Testing ────────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0
httpx>=0.27.0   # async test client for FastAPI
